Modifying structure for reading .properties files and separating the main.
